refactor(pruning): share tool result pruning with scroll
