+++ src/qwenpaw/agents/middlewares.py
@@ -16,8 +16,6 @@ Currently provided:
 
 import asyncio
 import logging
-import uuid
-from pathlib import Path
 from typing import TYPE_CHECKING, Any, AsyncGenerator, Callable, Set
 
 from agentscope.middleware import MiddlewareBase
@@ -25,14 +23,12 @@ from agentscope.message import Msg
 from agentscope.tool import ToolResponse
 
 from .tools.utils import (
-    truncate_text_output,
     DEFAULT_MAX_BYTES,
-    TRUNCATION_METADATA_KEY,
+    ToolResultPruner,
 )
 from ..constant import (
     AUTO_CONTINUE_MESSAGE_TAG,
     QWENPAW_MESSAGE_TAG_KEY,
-    TRUNCATION_NOTICE_MARKER,
 )
 
 if TYPE_CHECKING:
@@ -366,7 +362,7 @@ class ToolResultPruningMiddleware(MiddlewareBase):
         self._recent_max_bytes = recent_max_bytes
         self._exempt_extensions = exempt_file_extensions or set()
         self._exempt_tools = exempt_tool_names or set()
-        self._tool_results_dir = tool_results_dir
+        self._pruner = ToolResultPruner(tool_results_dir)
         self._agent_id = agent_id
 
     async def on_acting(
@@ -407,10 +403,10 @@ class ToolResultPruningMiddleware(MiddlewareBase):
         # ToolResponse byte size. Multi-block truncation metadata is kept by
         # content index so one block cannot influence another block's retry
         # location or cached file path.
-        self._prune_text_blocks(
+        self._pruner.prune_output(
             response.content or [],
-            self._recent_max_bytes,
-            response.metadata,
+            max_bytes=self._recent_max_bytes,
+            metadata=response.metadata,
         )
 
         return response
@@ -475,17 +471,15 @@ class ToolResultPruningMiddleware(MiddlewareBase):
                     if isinstance(block, dict)
                     else block.metadata
                 )
-                pruned, metadata = self._prune_output(
+                pruned, _ = self._pruner.prune_output(
                     output,
-                    effective_max,
-                    block_metadata,
+                    max_bytes=effective_max,
+                    metadata=block_metadata,
                 )
                 if isinstance(block, dict):
                     block["output"] = pruned
-                    block["metadata"].update(metadata)
                 else:
                     block.output = pruned
-                    block.metadata.update(metadata)
 
     def _detect_exempt_tool_ids(self, messages: list["Msg"]) -> Set[str]:
         exempt_ids: Set[str] = set()
@@ -538,138 +532,6 @@ class ToolResultPruningMiddleware(MiddlewareBase):
         if isinstance(block, dict):
             return block.get("type")
         return getattr(block, "type", None)
-
-    @staticmethod
-    def _block_text(block: Any) -> str:
-        if isinstance(block, dict):
-            return block.get("text", "")
-        return getattr(block, "text", "")
-
-    @staticmethod
-    def _set_block_text(block: Any, text: str) -> None:
-        if isinstance(block, dict):
-            block["text"] = text
-        else:
-            block.text = text
-
-    def _prune_text_blocks(
-        self,
-        blocks: list[Any],
-        max_bytes: int,
-        metadata: dict[str, Any],
-        encoding: str = "utf-8",
-    ) -> None:
-        """Prune text blocks while keeping metadata scoped to each block."""
-        text_indices = [
-            idx
-            for idx, block in enumerate(blocks)
-            if self._block_type(block) == "text"
-        ]
-        if not text_indices:
-            return
-
-        for idx in text_indices:
-            text = self._block_text(blocks[idx])
-            pruned, patch = self._truncate_tool_result(
-                text,
-                max_bytes,
-                metadata,
-                encoding,
-                block_index=idx,
-            )
-            self._set_block_text(blocks[idx], pruned)
-            patch_by_block = patch.get(TRUNCATION_METADATA_KEY)
-            if isinstance(patch_by_block, dict):
-                current = metadata.setdefault(TRUNCATION_METADATA_KEY, {})
-                if isinstance(current, dict):
-                    current.update(patch_by_block)
-
-    def _prune_output(
-        self,
-        output: str | list[Any],
-        max_bytes: int,
-        metadata: dict[str, Any],
-        encoding: str = "utf-8",
-    ) -> tuple[str | list[Any], dict[str, Any]]:
-        if isinstance(output, str):
-            return self._truncate_tool_result(
-                output,
-                max_bytes,
-                metadata,
-                encoding,
-            )
-        if isinstance(output, list):
-            self._prune_text_blocks(output, max_bytes, metadata, encoding)
-        return output, metadata
-
-    def _truncate_tool_result(
-        self,
-        content: str,
-        max_bytes: int,
-        metadata: dict[str, Any] | None = None,
-        encoding: str = "utf-8",
-        block_index: int = 0,
-    ) -> tuple[str, dict[str, Any]]:
-        if not content:
-            return content, {}
-
-        if TRUNCATION_NOTICE_MARKER in content:
-            return truncate_text_output(
-                content,
-                max_bytes=max_bytes,
-                metadata=metadata,
-                encoding=encoding,
-                block_index=block_index,
-            )
-
-        try:
-            content_bytes = len(content.encode(encoding))
-        except UnicodeEncodeError:
-            return content, {}
-
-        if content_bytes <= max_bytes:
-            return content, {}
-
-        saved_path = self._save_tool_result(content, encoding)
-        if saved_path is None:
-            return content, {}
-
-        return truncate_text_output(
-            content,
-            start_line=1,
-            total_lines=content.count("\n") + 1,
-            max_bytes=max_bytes,
-            file_path=saved_path,
-            encoding=encoding,
-            block_index=block_index,
-        )
-
-    def _save_tool_result(
-        self,
-        content: str,
-        encoding: str,
-    ) -> str | None:
-        """Persist a full tool result, returning ``None`` on failure."""
-        if not self._tool_results_dir:
-            logger.warning(
-                "Tool result exceeds the pruning limit but no artifact "
-                "directory is configured; returning the original result",
-            )
-            return None
-
-        try:
-            tool_result_dir = Path(self._tool_results_dir)
-            tool_result_dir.mkdir(parents=True, exist_ok=True)
-            fp = tool_result_dir / f"{uuid.uuid4().hex}.txt"
-            fp.write_text(content, encoding=encoding)
-        except OSError as e:
-            logger.warning(
-                "Failed to save tool result to file; returning the original "
-                "result: %s",
-                e,
-            )
-            return None
-        return str(fp)
 
 
 class LangfuseToolSpanMiddleware(MiddlewareBase):
